[wip] axi_test (rand_axi_lite_slave): Specify default values for parameters
TODO: Set the default value to 0.  This will disable the VCS workaround,
because VCS fails to correctly pass the value of the parameter.
However, if we do not set the default value to 0, VCS will always take
the default value and not complain about the mismatch.  Having a hidden
value that is used instead of the passed parameter in a module/class
instantiated in multiple places is worse than VCS failing to compile,
because it can lead to mismatches that are difficult to debug.